Rename duckdb/utils.py to duckdb/schema_inspection.py
