+++ databao/duckdb/__init__.py
@@ -1,5 +1,5 @@
 from databao.duckdb.react_tools import AgentResponse, make_duckdb_tool, make_react_duckdb_agent
-from databao.duckdb.utils import describe_duckdb_schema
+from databao.duckdb.schema_inspection import describe_duckdb_schema
 
 __all__ = [
     "AgentResponse",
