+++ databao/duckdb/schema_inspection.py
